Replace Akka with Apache Pekko
Replace all Akka modules with Apache Pekko to avoid Akka's BSL
licensing concerns while maintaining the same streaming capabilities.

## Changes

**Modules:**
- Remove: alternator-akka-base, alternator-akka-aws1, alternator-akka-aws2
- Add: alternator-pekko-base, alternator-pekko-aws1, alternator-pekko-aws2

**Implementation:**
- Apache Pekko 1.1.2 (API-compatible fork of Akka 2.6.21)
- All imports changed from `akka.*` to `org.apache.pekko.*`
- Classes renamed: AkkaAws1/2 → PekkoAws1/2, AkkaBase → PekkoBase
- Scala 2.12: Custom parasitic ExecutionContext implementation
- Scala 2.13: Uses stdlib ExecutionContext.parasitic

**Testing:**
- All 120 integration tests passing (60 per AWS SDK version)
- Behavioral equivalence with previous Akka implementation verified

**Documentation:**
- Updated README.md with Pekko examples
- Updated CLAUDE.md with new module architecture
- Removed Akka-to-Pekko migration guide (no longer needed)

## Breaking Changes

This is a breaking change for users of Akka modules. To migrate:
1. Replace `alternator-akka-aws*` with `alternator-pekko-aws*`
2. Update imports: `akka.*` → `org.apache.pekko.*`
3. Update instantiations: `AkkaAws1/2()` → `PekkoAws1/2()`
