--- akka-base/src/main/scala-2.13/com/hiya/alternator/akka/JdkCompat.scala
@@ -1,12 +1,0 @@
-package com.hiya.alternator.akka
-
-import java.util.concurrent.CompletableFuture
-import scala.concurrent.{ExecutionContext, Future}
-
-private[akka] object JdkCompat {
-  implicit lazy val parasitic: ExecutionContext = ExecutionContext.parasitic
-
-  implicit class CompletionStage[T](val cs: CompletableFuture[T]) extends AnyVal {
-    @inline def asScala: Future[T] = scala.jdk.FutureConverters.CompletionStageOps(cs).asScala
-  }
-}
